+++ packages/blockstore-core/src/memory.ts
@@ -5,7 +5,7 @@ import * as raw from 'multiformats/codecs/raw'
 import * as Digest from 'multiformats/hashes/digest'
 import { BaseBlockstore } from './base.js'
 import type { Pair } from 'interface-blockstore'
-import type { Await, AwaitIterable } from 'interface-store'
+import type { AbortOptions, Await, AwaitIterable } from 'interface-store'
 
 export class MemoryBlockstore extends BaseBlockstore {
   private readonly data: Map<string, Uint8Array>
@@ -16,13 +16,15 @@ export class MemoryBlockstore extends BaseBlockstore {
     this.data = new Map()
   }
 
-  put (key: CID, val: Uint8Array): Await<CID> { // eslint-disable-line require-await
+  put (key: CID, val: Uint8Array, options?: AbortOptions): Await<CID> {
+    options?.signal?.throwIfAborted()
     this.data.set(base32.encode(key.multihash.bytes), val)
 
     return key
   }
 
-  get (key: CID): Await<Uint8Array> {
+  get (key: CID, options?: AbortOptions): Await<Uint8Array> {
+    options?.signal?.throwIfAborted()
     const buf = this.data.get(base32.encode(key.multihash.bytes))
 
     if (buf == null) {
@@ -32,20 +34,25 @@ export class MemoryBlockstore extends BaseBlockstore {
     return buf
   }
 
-  has (key: CID): Await<boolean> {
+  has (key: CID, options?: AbortOptions): Await<boolean> {
+    options?.signal?.throwIfAborted()
     return this.data.has(base32.encode(key.multihash.bytes))
   }
 
-  async delete (key: CID): Promise<void> {
+  async delete (key: CID, options?: AbortOptions): Promise<void> {
+    options?.signal?.throwIfAborted()
     this.data.delete(base32.encode(key.multihash.bytes))
   }
 
-  async * getAll (): AwaitIterable<Pair> {
+  async * getAll (options?: AbortOptions): AwaitIterable<Pair> {
+    options?.signal?.throwIfAborted()
+
     for (const [key, value] of this.data.entries()) {
       yield {
         cid: CID.createV1(raw.code, Digest.decode(base32.decode(key))),
         block: value
       }
+      options?.signal?.throwIfAborted()
     }
   }
 }