+++ packages/blockstore-fs/test/fixtures/writer-worker.ts
@@ -1,10 +1,11 @@
 import { CID } from 'multiformats/cid'
+// @ts-expect-error types are broken: https://github.com/andywer/threads.js/pull/470
 import { expose } from 'threads/worker'
 import { FsBlockstore } from '../../src/index.js'
 
 let fs: FsBlockstore
 expose({
-  async isReady (path) {
+  async isReady (path: string) {
     fs = new FsBlockstore(path)
     try {
       await fs.open()
@@ -15,7 +16,7 @@ expose({
       throw err
     }
   },
-  async put (cidString, value) {
+  async put (cidString: string, value: Uint8Array) {
     const key = CID.parse(cidString)
     try {
       return await fs.put(key, value)