+++ packages/blockstore-fs/test/index.spec.ts
@@ -1,3 +1,4 @@
+import { setMaxListeners } from 'node:events'
 import fs from 'node:fs/promises'
 import os from 'node:os'
 import path from 'node:path'
@@ -5,6 +6,7 @@ import { expect } from 'aegir/chai'
 import { interfaceBlockstoreTests } from 'interface-blockstore-tests'
 import { base32 } from 'multiformats/bases/base32'
 import { CID } from 'multiformats/cid'
+// @ts-expect-error types are broken: https://github.com/andywer/threads.js/pull/470
 import { spawn, Thread, Worker } from 'threads'
 import { FsBlockstore } from '../src/index.js'
 import { FlatDirectory, NextToLast } from '../src/sharding.js'
@@ -97,7 +99,10 @@ describe('FsBlockstore', () => {
       teardown: async (store) => {
         await store.close()
         await fs.rm(store.path, {
-          recursive: true
+          recursive: true,
+          force: true,
+          maxRetries: 5,
+          retryDelay: 1_000
         })
       }
     })
@@ -114,7 +119,10 @@ describe('FsBlockstore', () => {
       teardown: async (store) => {
         await store.close()
         await fs.rm(store.path, {
-          recursive: true
+          recursive: true,
+          force: true,
+          maxRetries: 5,
+          retryDelay: 1_000
         })
       }
     })
@@ -136,7 +144,10 @@ describe('FsBlockstore', () => {
       teardown: async (store) => {
         await store.close()
         await fs.rm(store.path, {
-          recursive: true
+          recursive: true,
+          force: true,
+          maxRetries: 5,
+          retryDelay: 1_000
         })
       }
     })
@@ -168,7 +179,9 @@ describe('FsBlockstore', () => {
     const dir = path.join(os.tmpdir(), `test-${Math.random()}`)
     const key = CID.parse('QmeimKZyjcBnuXmAD9zMnSjM9JodTbgGT3gutofkTqz9rE')
     const workers = await Promise.all(new Array(10).fill(0).map(async () => {
-      const worker = await spawn(new Worker('./fixtures/writer-worker.js'))
+      const w = new Worker('./fixtures/writer-worker.js')
+      setMaxListeners(Infinity, w)
+      const worker = await spawn(w)
       await worker.isReady(dir)
       return worker
     }))
@@ -177,7 +190,7 @@ describe('FsBlockstore', () => {
       const value = utf8Encoder.encode('Hello world')
       // 100 iterations of looping over all workers and putting the same key value pair
       await Promise.all(new Array(100).fill(0).map(async () => {
-        return Promise.all(workers.map(async (worker) => worker.put(key.toString(), value)))
+        return Promise.all(workers.map(async (worker: any) => worker.put(key.toString(), value)))
       }))
 
       const fs = new FsBlockstore(dir)
@@ -186,7 +199,7 @@ describe('FsBlockstore', () => {
 
       expect(res).to.deep.equal(value)
     } finally {
-      await Promise.all(workers.map(async (worker) => Thread.terminate(worker)))
+      await Promise.all(workers.map(async (worker: any) => Thread.terminate(worker)))
     }
   })
 })