+++ packages/datastore-core/src/black-hole.ts
@@ -1,31 +1,36 @@
 import { NotFoundError } from 'interface-store'
 import { BaseDatastore } from './base.js'
-import type { Pair } from 'interface-datastore'
+import type { Pair, Query, KeyQuery } from 'interface-datastore'
 import type { Key } from 'interface-datastore/key'
-import type { Await, AwaitIterable } from 'interface-store'
+import type { AbortOptions, Await, AwaitIterable } from 'interface-store'
 
 export class BlackHoleDatastore extends BaseDatastore {
-  put (key: Key): Await<Key> {
+  put (key: Key, value: Uint8Array, options?: AbortOptions): Await<Key> {
+    options?.signal?.throwIfAborted()
     return key
   }
 
-  get (): Await<Uint8Array> {
+  get (key: Key, options?: AbortOptions): Await<Uint8Array> {
+    options?.signal?.throwIfAborted()
     throw new NotFoundError()
   }
 
-  has (key: Key): Await<boolean> {
+  has (key: Key, options?: AbortOptions): Await<boolean> {
+    options?.signal?.throwIfAborted()
     return false
   }
 
-  delete (key: Key): Await<void> {
-
+  delete (key: Key, options?: AbortOptions): Await<void> {
+    options?.signal?.throwIfAborted()
   }
 
-  * _all (): AwaitIterable<Pair> {
-
+  // eslint-disable-next-line require-yield
+  * _all (q: Query, options?: AbortOptions): AwaitIterable<Pair> {
+    options?.signal?.throwIfAborted()
   }
 
-  * _allKeys (): AwaitIterable<Key> {
-
+  // eslint-disable-next-line require-yield
+  * _allKeys (q: KeyQuery, options?: AbortOptions): AwaitIterable<Key> {
+    options?.signal?.throwIfAborted()
   }
 }