+++ packages/datastore-core/src/memory.ts
@@ -1,8 +1,8 @@
 import { Key } from 'interface-datastore/key'
 import { NotFoundError } from 'interface-store'
 import { BaseDatastore } from './base.js'
-import type { Pair } from 'interface-datastore'
-import type { Await, AwaitIterable } from 'interface-store'
+import type { KeyQuery, Pair, Query } from 'interface-datastore'
+import type { AbortOptions, Await, AwaitIterable } from 'interface-store'
 
 export class MemoryDatastore extends BaseDatastore {
   private readonly data: Map<string, Uint8Array>
@@ -13,13 +13,17 @@ export class MemoryDatastore extends BaseDatastore {
     this.data = new Map()
   }
 
-  put (key: Key, val: Uint8Array): Await<Key> { // eslint-disable-line require-await
+  put (key: Key, val: Uint8Array, options?: AbortOptions): Await<Key> {
+    options?.signal?.throwIfAborted()
+
     this.data.set(key.toString(), val)
 
     return key
   }
 
-  get (key: Key): Await<Uint8Array> {
+  get (key: Key, options?: AbortOptions): Await<Uint8Array> {
+    options?.signal?.throwIfAborted()
+
     const result = this.data.get(key.toString())
 
     if (result == null) {
@@ -29,23 +33,29 @@ export class MemoryDatastore extends BaseDatastore {
     return result
   }
 
-  has (key: Key): Await<boolean> { // eslint-disable-line require-await
+  has (key: Key, options?: AbortOptions): Await<boolean> {
+    options?.signal?.throwIfAborted()
     return this.data.has(key.toString())
   }
 
-  delete (key: Key): Await<void> { // eslint-disable-line require-await
+  delete (key: Key, options?: AbortOptions): Await<void> {
+    options?.signal?.throwIfAborted()
     this.data.delete(key.toString())
   }
 
-  * _all (): AwaitIterable<Pair> {
+  * _all (q: Query, options?: AbortOptions): AwaitIterable<Pair> {
+    options?.signal?.throwIfAborted()
     for (const [key, value] of this.data.entries()) {
       yield { key: new Key(key), value }
+      options?.signal?.throwIfAborted()
     }
   }
 
-  * _allKeys (): AwaitIterable<Key> {
+  * _allKeys (q: KeyQuery, options?: AbortOptions): AwaitIterable<Key> {
+    options?.signal?.throwIfAborted()
     for (const key of this.data.keys()) {
       yield new Key(key)
+      options?.signal?.throwIfAborted()
     }
   }
 }