+++ packages/datastore-s3/CHANGELOG.md
@@ -1,3 +1,16 @@
+## [datastore-s3-v12.0.4](https://github.com/ipfs/js-stores/compare/datastore-s3-12.0.3...datastore-s3-12.0.4) (2025-05-28)
+
+### Bug Fixes
+
+* improve abort signal support ([#350](https://github.com/ipfs/js-stores/issues/350)) ([e17d770](https://github.com/ipfs/js-stores/commit/e17d770cc2fcee77cb0152a855abf162e5a91a99))
+
+## [datastore-s3-v12.0.3](https://github.com/ipfs/js-stores/compare/datastore-s3-12.0.2...datastore-s3-12.0.3) (2025-05-26)
+
+### Dependencies
+
+* bump aegir from 44.1.4 to 47.0.16 ([#349](https://github.com/ipfs/js-stores/issues/349)) ([d33d15f](https://github.com/ipfs/js-stores/commit/d33d15f0638856530d0e1868c723e5567abf27e6))
+* **dev:** bump sinon from 19.0.5 to 20.0.0 ([#337](https://github.com/ipfs/js-stores/issues/337)) ([ec2a54a](https://github.com/ipfs/js-stores/commit/ec2a54a1ef2ecbee862d17710430a51b79063183))
+
 ## [datastore-s3-v12.0.2](https://github.com/ipfs/js-stores/compare/datastore-s3-12.0.1...datastore-s3-12.0.2) (2024-09-13)
 
 ### Dependencies