+++ packages/interface-blockstore/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [interface-blockstore-v5.3.2](https://github.com/ipfs/js-stores/compare/interface-blockstore-5.3.1...interface-blockstore-5.3.2) (2025-05-26)
+
+### Dependencies
+
+* bump aegir from 44.1.4 to 47.0.16 ([#349](https://github.com/ipfs/js-stores/issues/349)) ([d33d15f](https://github.com/ipfs/js-stores/commit/d33d15f0638856530d0e1868c723e5567abf27e6))
+
 ## [interface-blockstore-v5.3.1](https://github.com/ipfs/js-stores/compare/interface-blockstore-5.3.0...interface-blockstore-5.3.1) (2024-09-13)
 
 ### Bug Fixes