+++ packages/interface-datastore-tests/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [interface-datastore-tests-v6.0.3](https://github.com/ipfs/js-stores/compare/interface-datastore-tests-6.0.2...interface-datastore-tests-6.0.3) (2025-05-28)
+
+### Bug Fixes
+
+* improve abort signal support ([#350](https://github.com/ipfs/js-stores/issues/350)) ([e17d770](https://github.com/ipfs/js-stores/commit/e17d770cc2fcee77cb0152a855abf162e5a91a99))
+
+## [interface-datastore-tests-v6.0.2](https://github.com/ipfs/js-stores/compare/interface-datastore-tests-6.0.1...interface-datastore-tests-6.0.2) (2025-05-26)
+
+### Dependencies
+
+* bump aegir from 44.1.4 to 47.0.16 ([#349](https://github.com/ipfs/js-stores/issues/349)) ([d33d15f](https://github.com/ipfs/js-stores/commit/d33d15f0638856530d0e1868c723e5567abf27e6))
+
 ## [interface-datastore-tests-v6.0.1](https://github.com/ipfs/js-stores/compare/interface-datastore-tests-6.0.0...interface-datastore-tests-6.0.1) (2024-09-13)
 
 ### Bug Fixes